+++ ext.php
@@ -25,12 +25,12 @@ class ext extends \phpbb\extension\base
 	 * The current phpBB version should meet or exceed
 	 * the minimum version required by this extension:
 	 *
-	 * Requires phpBB 3.3.0 and PHP 7.1
+	 * Requires phpBB 3.3.18 and PHP 7.1
 	 *
 	 * @return bool
 	 */
 	public function is_enableable()
 	{
-		return phpbb_version_compare(PHPBB_VERSION, '3.3.0', '>=') && PHP_VERSION_ID >= 70100;
+		return phpbb_version_compare(PHPBB_VERSION, '3.3.18-dev', '>=') && PHP_VERSION_ID >= 70100;
 	}
 }